Move AuthAccessTokenFetcher to core/modules/oauth
