+++ core/src/index.ts
@@ -55,6 +55,8 @@ export { AppsCache } from "./modules/caches/appsCache";
 export { SourcesCache } from "./modules/caches/sourcesCache";
 export { PropertiesCache } from "./modules/caches/propertiesCache";
 
+export { oAuthAccessTokenFetcher } from "./modules/oAuth";
+
 export { Status } from "./modules/status";
 export { StatusMetric } from "./modules/statusReporters";
 export { GrouparooCLI } from "./modules/cli";
